Clean Result View Controller and set constraints for elements
