+++ Who-wanna-be-millionaire/Application/SceneDelegate.swift
@@ -21,7 +21,8 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         
         window = UIWindow(frame: windowScene.coordinateSpace.bounds)
         window?.windowScene = windowScene
-        window?.rootViewController = WelcomeViewController()
+        // Change it back after module will be completed
+        window?.rootViewController = ResultViewController()
         window?.makeKeyAndVisible()
     }
